requirements.txt
python-telegram-bot[webhooks]==21.0.1
APScheduler==3.10.4
python-dotenv==1.0.0
Flask==3.0.0
gunicorn==21.2.0
requests==2.31.0
gspread==5.12.4
google-auth-oauthlib==1.2.0
pytz==2023.3.post1
